Use custom create token
