+++ mpi4jax/__init__.py
@@ -2,6 +2,8 @@ import mpi4jax.cython  # noqa: F401
 
 __all__ = ["Allreduce", "Send", "Recv", "Sendrecv"]
 
+from ._create_token import create_token
+
 from .collective_ops.allreduce import Allreduce
 from .collective_ops.send import Send
 from .collective_ops.recv import Recv
