+++ mpi4jax/collective_ops/allreduce.py
@@ -4,11 +4,11 @@ from mpi4py import MPI as _MPI
 
 from jax import abstract_arrays, device_put
 from jax import numpy as jnp
-from jax.lax import create_token, zeros_like_array
-from jax.lax.lax import create_token_p
 from jax.core import Primitive
 from jax.lib import xla_client
 from jax.interpreters import xla, ad
+
+from .. import create_token
 
 from ..utils import (
     to_mpi_ptr,
@@ -138,18 +138,3 @@ ad.primitive_jvps[mpi_allreduce_p] = mpi_allreduce_value_and_jvp
 
 # assign to the primitive the correct encoder
 xla.backend_specific_translations["cpu"][mpi_allreduce_p] = mpi_allreduce_xla_encode
-
-
-# This is a fix to the fact that `jax.lax.create_token` does not have an
-# adjoint defined, so we define it here.
-# Hopefully this issue will be resolved in upstream at a later date
-def create_token_value_and_jvp(in_args, tan_args):
-    print("create in; ", in_args)
-    print("create tan_in: ", tan_args)
-    (x,) = in_args
-    res = create_token(x)
-    jvp = zeros_like_array(x)
-    return (res, jvp)
-
-
-ad.primitive_jvps[create_token_p] = create_token_value_and_jvp
